Removing the /.. from the browser path.
If the parent folder is selected (..) when the SELECT button is pressed,
we understand that user wants to explore the current folder.

This assertion is done in the gui/form.py when the browser is launched and
the path is retrieved. In the gui/browser.py (FileBrowser class) is not clear
how to do it.


Former-commit-id: 0bfb410527a032eff67f5f929657a04b4179bd26